refs: dat_phong_client

+++ database/migrations/2024_03_18_081650_create_chi_tiet_dat_phongs_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->unsignedBigInteger('dat_phong_id');
             $table->foreign('dat_phong_id')->references('id')->on('dat_phongs');
-            $table->decimal('thanh_tien')->nullable();
+            $table->float('thanh_tien')->nullable();
             $table->softDeletes();
             $table->timestamps();
         });
